Do not allow the helper to set duplicate public keys
